Zach's comments
