--- sql/information_schema/statistics_table.go
@@ -1,1 +1,0 @@
-package information_schema
